refactor: Rename dflowd to chatsky-ui
