+++ backend/df_designer/chatsky_ui/api/api_v1/endpoints/config.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+from chatsky_ui import __version__
+
+router = APIRouter()
+
+
+@router.get("/version")
+async def get_version():
+    return __version__
