--- backend/df_designer/dflowd/api/api_v1/endpoints/config.py
@@ -1,10 +1,0 @@
-from fastapi import APIRouter
-
-from dflowd import __version__
-
-router = APIRouter()
-
-
-@router.get("/version")
-async def get_version():
-    return __version__
