--- backend/df_designer/dflowd/api/deps.py
@@ -1,30 +1,0 @@
-from dflowd.services.index import Index
-from dflowd.services.process_manager import BuildManager, RunManager
-from dflowd.services.websocket_manager import WebSocketManager
-
-build_manager = BuildManager()
-
-
-def get_build_manager() -> BuildManager:
-    return build_manager
-
-
-run_manager = RunManager()
-
-
-def get_run_manager() -> RunManager:
-    return run_manager
-
-
-websocket_manager = WebSocketManager()
-
-
-def get_websocket_manager() -> WebSocketManager:
-    return websocket_manager
-
-
-index = Index()
-
-
-def get_index() -> Index:
-    return index
