+++ backend/df_designer/pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "dflowd"
+name = "chatsky-ui"
 version = "0.1.0b4"
-description = "Dialog Flow Designer"
+description = "Chatsky-UI is GUI for Chatsky Framework, that is a free and open-source software stack for creating chatbots, released under the terms of Apache License 2.0."
 license = "Apache-2.0"
 authors = [
     "Denis Kuznetsov <kuznetsov.den.p@gmail.com>",
@@ -9,7 +9,7 @@ authors = [
     "Rami Mashkouk <rami.n.mashkouk@gmail.com>",
 ]
 readme = "README.md"
-packages = [{include = "dflowd"}]
+packages = [{include = "chatsky_ui"}]
 
 [tool.poetry.dependencies]
 python = "^3.8.1"
@@ -32,7 +32,7 @@ sphinx = { version = "*", python = "^3.10"}
 sphinx-rtd-theme = { version = "*", python = "^3.10"}
 
 [tool.poetry.scripts]
-dflowd = "dflowd.cli:cli"
+"chatsky.ui" = "chatsky_ui.cli:cli"
 
 [tool.poetry.group.lint]
 optional = true
